crete registration controller and email service
