+++ flashcards-api/src/main/java/com/chornobuk/flashcardsapi/utils/ConfirmationLetterBuilder.java
@@ -1,0 +1,4 @@
+package com.chornobuk.flashcardsapi.utils;
+
+public class ConfirmationLetterBuilder {
+}
